# server/docs/swagger.yaml
basePath: /
definitions:
...
  dtos.EventDTO:
    properties:
      created_by:
        type: string
      description:
        minLength: 3
        type: string
      end_date:
        type: string
      id:
        type: string
      is_private:
        type: boolean
      participants:
        items:
          $ref: '#/definitions/dtos.ParticipantDTO'
        type: array
      physical_event:
        $ref: '#/definitions/dtos.PhysicalEventDTO'
      remote_event:
        $ref: '#/definitions/dtos.RemoteEventDTO'
      start_date:
        type: string
      subjects:
        items:
          type: string
        type: array
      title:
        minLength: 3
        type: string
    required:
    - description
    - end_date
    - start_date
    - title
    type: object
...
  dtos.ParticipantDTO:
    properties:
      event_id:
        type: string
      id:
        type: string
      joined_at:
        type: string
      requested_at:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  dtos.PhysicalEventDTO:
    properties:
      id:
        type: string
      lat:
        type: number
      lng:
        type: number
      location:
        type: string
    type: object
...
  dtos.RemoteEventDTO:
    properties:
      id:
        type: string
      url:
        type: string
    required:
    - url
    type: object
...
  dtos.UserDTO:
    properties:
      activated:
        type: boolean
      address:
        type: string
      bio:
        type: string
      birthDate:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      picture:
        type: string
      reportNumber:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/user.Role'
        enum:
        - SUPERADMIN
        - ADMIN
        - USER
      username:
        minLength: 3
        type: string
    required:
    - email
    - id
    - role
    - username
    type: object
...
  user.Role:
    enum:
    - USER
    - SUPER ADMIN
    - ADMIN
    - USER
    type: string
    x-enum-varnames:
    - DefaultRole
    - RoleSUPERADMIN
    - RoleADMIN
    - RoleUSER
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the EduMeet project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: EduMeet API Documentation
  version: "1.0"
paths:
...
  /events:
    get:
      consumes:
      - application/json
      description: Get a list of all events with optional filters
      parameters:
      - default: '"all"'
        description: Event Type
        in: query
        name: type
        type: string
      - description: Distance Filter
        in: query
        name: distance
        type: string
      - description: Longitude Filter
        in: query
        name: longitude
        type: string
      - description: Latitude Filter
        in: query
        name: latitude
        type: string
      - description: Subjects Filter
        in: query
        name: subjects
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of Events
          schema:
            items:
              $ref: '#/definitions/dtos.EventDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all events
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Create a new event with the given details
      parameters:
      - description: Event Details
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/dtos.EventDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Event Created Successfully
          schema:
            $ref: '#/definitions/dtos.EventDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Validation Errors
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new event
      tags:
      - Events
  /events/{eventID}/participants/pending:
    get:
      consumes:
      - application/json
      description: Get a list of participants who are pending approval for an event
      parameters:
      - description: Event ID
        in: path
        name: eventID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of Pending Participants
          schema:
            items:
              $ref: '#/definitions/dtos.UserDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get pending participants for an event
      tags:
      - Events
  /events/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an event based on the provided ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 'Success: No Content'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an event
      tags:
      - Events
    get:
      consumes:
      - application/json
      description: Get an event's details based on the provided ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event Found
          schema:
            $ref: '#/definitions/dtos.EventDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an event
      tags:
      - Events
    put:
      consumes:
      - application/json
      description: Update the details of an event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Event Details
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/dtos.EventDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Event Updated Successfully
          schema:
            $ref: '#/definitions/dtos.EventDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Validation Errors
          schema:
            additionalProperties: true
            type: object
      summary: Update an event
      tags:
      - Events
  /events/{id}/details:
    get:
      consumes:
      - application/json
      description: Get event details by ID including additional information
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event Details
          schema:
            $ref: '#/definitions/dtos.EventDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get event with details
      tags:
      - Events
  /events/code/{eventId}:
    get:
      consumes:
      - application/json
      description: Retrieves the event code required to join the event
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event Code
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get event code
      tags:
      - Events
  /events/created-by/current:
    get:
      consumes:
      - application/json
      description: Get a list of events created by the current user
      produces:
      - application/json
      responses:
        "200":
          description: List of Created Events
          schema:
            items:
              $ref: '#/definitions/dtos.EventDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get events created by the current user
      tags:
      - Events
  /events/join/{code}:
    get:
      consumes:
      - application/json
      description: Allows a user to join an event by using the provided join code
      parameters:
      - description: Join Code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully Joined Event
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Join an event by code
      tags:
      - Events
  /events/subjects/update/{id}:
    put:
      consumes:
      - application/json
      description: Update the list of subjects associated with the event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: List of Subjects
        in: body
        name: subjects
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Subjects Updated Successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update event subjects
      tags:
      - Events
  /events/users/current:
    get:
      consumes:
      - application/json
      description: Get a list of events associated with the current user
      produces:
      - application/json
      responses:
        "200":
          description: List of Current User's Events
          schema:
            items:
              $ref: '#/definitions/dtos.EventDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current user's events
      tags:
      - Events
...
swagger: "2.0"